added query, rules and users class
