+++ irodsConnector/__init__.py
@@ -7,7 +7,9 @@ from . import keywords
 from . import manager
 from . import meta
 from . import permission
+from . import query
+from . import rules
 from . import resource
 from . import session
 from . import tickets
-from . import utils
+from . import users
