+++ irodsConnector/query.py
@@ -1,45 +1,11 @@
-""" irods utils
+""" query operations
 """
-import logging
-import irods.collection
-import irods.data_object
-import irods.exception
-import irods.rule
 import irodsConnector.keywords as kw
 from irodsConnector.session import Session
 
 
-class IrodsUtils(object):
-    """Irods calls which don't fit in one of the existing groups"""
-
-    def get_user_info(self, ses_man: Session) -> tuple:
-        """Query for user type and groups.
-
-        Parameters
-        ----------
-        ses_man : irods session
-            Instance of the Session class
-
-        Returns
-        -------
-        list
-            iRODS user type names
-        list
-            iRODS group names
-
-        """
-        query = ses_man.session.query(kw.USER_TYPE).filter(kw.LIKE(
-            kw.USER_NAME, ses_man.session.username))
-        user_type = [
-            list(result.values())[0] for result in query.get_results()
-        ][0]
-        query = ses_man.session.query(kw.USER_GROUP_NAME).filter(kw.LIKE(
-            kw.USER_NAME, ses_man.session.username))
-        user_groups = [
-            list(result.values())[0] for result in query.get_results()
-        ]
-        return user_type, user_groups
-
+class Query(object):
+    """Irods Query operations """
     def search(self, ses_man: Session, key_vals: dict = None) -> list:
         """Given a dictionary with metadata attribute names as keys and
         associated values, query for collections and data objects that
@@ -128,53 +94,3 @@ class IrodsUtils(object):
                             res[list(res.keys())[1]],
                             res[list(res.keys())[2]]])
         return results
-
-    def execute_rule(self, ses_man: Session, rule_file: str, params: dict, output: str = 'ruleExecOut') -> tuple:
-        """Execute an iRODS rule.
-
-        Parameters
-        ----------
-        ses_man : irods session
-            Instance of the Session class
-        rule_file : str, file-like
-            Name of the iRODS rule file, or a file-like object representing it.
-        params : dict
-            Rule arguments.
-        output : str
-            Rule output variable(s).
-
-        Returns
-        -------
-        tuple
-            (stdout, stderr)
-
-        `params` format example:
-        params = {  # extra quotes for string literals
-            '*obj': '"/zone/home/user"',
-            '*name': '"attr_name"',
-            '*value': '"attr_value"'
-        }
-
-        """
-        try:
-            rule = irods.rule.Rule(
-                ses_man.session, rule_file=rule_file, params=params, output=output,
-                instance_name='irods_rule_engine_plugin-irods_rule_language-instance')
-            out = rule.execute()
-        except irods.exception.NetworkException as netexc:
-            logging.info('Lost connection to iRODS server.')
-            return '', repr(netexc)
-        except irods.exception.SYS_HEADER_READ_LEN_ERR as shrle:
-            logging.info('iRODS server hiccuped.  Check the results and try again.')
-            return '', repr(shrle)
-        except Exception as error:
-            logging.info('RULE EXECUTION ERROR', exc_info=True)
-            return '', repr(error)
-        stdout, stderr = '', ''
-        if len(out.MsParam_PI) > 0:
-            buffers = out.MsParam_PI[0].inOutStruct
-            stdout = (buffers.stdoutBuf.buf or b'').decode()
-            # Remove garbage after terminal newline.
-            stdout = '\n'.join(stdout.split('\n')[:-1])
-            stderr = (buffers.stderrBuf.buf or b'').decode()
-        return stdout, stderr
